Refactor `WebhooksService#ping` to provide access to the raw API response where successful
The "Ping a webhook" action in the Duffel API is unlike any other
action - it returns a `204 No Content` response if successful. This
unsurprisingly means it ends up being a bit of an exception in our
code!

Right now, it returns `true` if the ping was successful or
otherwise raises an error. This means that you can't get access
to the raw API response in the success case - which is available
for all other actions (see #5).

This changes `#ping` to return an instance of a new class called
`PingResult` - it confors to the API of other results in the
library thanks to its `#api_response` public method. It also
exposes a `#succeeded` method, which returns `true`.

Failure cases continue to raise an error.