Rename `penalty_amount` and `penalty_currency` across the library to `penalty_total_amount` and `penalty_total_currency`
The client library was written based on our API reference. It
turns out that it mentions `penalty_amount` and `penalty_currency`
attributes in the API in a bunch of places, that actually have
`_total_` in their names in the serializers.

This fixes that so the data is exposed on the relevant resources.
I've left `TODO`s behind to fix the issues with the docs.