Rationalise the interface of `DuffelAPI::APIResponse`
This commit rationalises the public interface of `DuffelAPI::
APIResponse` which we use to expose "raw API responses" across
the library (i.e. provide access to the status code, body,
headers and request ID).

It has a few extra public methods at the moment that aren't
really necessarily and I'd prefer to keep it small and *tight*.
We can always add more, but it's hard to take away!

Note that `APIResponse` is different to `Response` which is the
representation of a raw API response used internally inside the
library.