Don't include a `code` when raising an error for non-standard errors
Our standard `code`s on our errors are `String`s, but we are
currently using the HTTP status code, an intgeer, for the
`code` for non-standard errors (e.g. those returned by a load
balancer). This stops doing that, preferring `nil`, to avoid
having a weird might-be-integer-or-might-be-string API for
`Error#code`.