Allow verifying the genuineness of webhook events with `WebhookEvent.genuine?`
This adds a new method, `WebhookEvent.genuine?`, which can check
that a webhook event received from Duffel is genuinely from
Duffel. We're able to do this because the Duffel API signs
webhook events with a secret known only to the integrator and
Duffel.

This is based on Python code in our [guides][1] and tested
against real example webhooks, just to make sure it works!

[1]: https://duffel.com/docs/guides/receiving-webhooks