Add GitHub Actions workflow for publishing gem versions
Building and pushing gems to RubyGems is tedious work but a small error
while performing the steps involved can have unintended consequences.
Using a GitHub Action to automate them should simplify the process of
pushing new gems and enable anyone with the ability to push tags to this
repository to trigger a run of this workflow.