Move Code Examples to a different workflow
Code Examples job always fails because it requires secrets that
are not available for external contributors. So the idea here is
move that job from the normal CI and only make it run on main. It
will simplify external contribution while keeping running the code
examples where it matters the most, on main.